removed unless components and imported to TicketOrder.vue
